Set up react browser router and move files
