--- client/src/CardsPage/CardsPage.jsx
@@ -1,8 +1,0 @@
-function CardsPage() {
-    return (
-        <div>
-        <h1>Cards Page</h1>
-        </div>
-    );
-}
-export default CardsPage;
