+++ client/src/pages/HomePage/HomePage.jsx
@@ -6,8 +6,6 @@ import propTypes from "prop-types";
 HomePage.propTypes = {
     userData: propTypes.array.isRequired,
     boardData: propTypes.array.isRequired,
-    setInHome: propTypes.func.isRequired,
-    setViewingBoardId: propTypes.func.isRequired,
 }
 
 function HomePage(props) {
@@ -40,8 +38,6 @@ function HomePage(props) {
                             )}
                             boardDataFunc={props.boardData[1]}
                             filter={filter}
-                            setInHome={props.setInHome}
-                            setViewingBoardId={props.setViewingBoardId}
                         />
                     );
                 })}
