+++ client/src/pages/HomePage/components/BoardCard.jsx
@@ -1,31 +1,24 @@
 import propTypes from "prop-types";
-import API from "../../api";
+import API from "../../../api";
 import "./BoardCard.css";
 
 BoardCard.propTypes = {
     userData: propTypes.array.isRequired,
     boardData: propTypes.array.isRequired,
+
     board: propTypes.object.isRequired,
     isOwned: propTypes.bool.isRequired, // changes the ui to allow deletion and stops self-upvotes
     boardDataFunc: propTypes.func.isRequired,
     filter: propTypes.string.isRequired,
-    setInHome: propTypes.func.isRequired,
-    setViewingBoardId: propTypes.func.isRequired,
 };
 
 function BoardCard(props) {
     // handle the deletion of a board card
-    const handleDeleteBoardCard = (event) => {
+    const handleDeleteBoardCard = () => {
         API.deleteBoard(() => {}, props.board["id"], props.userData["id"]);
         // delete from boardData state
         props.boardData[1](props.boardData[0].filter((board) => board["id"] !== props.board["id"]));
-    }
-
-    // handle the viewing of a board
-    const handleViewBoard = (event) => {
-        props.setInHome(false);
-        props.setViewingBoardId(props.board["id"]);
-    } 
+    };
 
     return (
         <div
@@ -50,13 +43,13 @@ function BoardCard(props) {
                         <div className="d-flex align-items-center">
                             <i className="bi bi-trash-fill h5 m-0 trash" onClick={handleDeleteBoardCard}></i>
                         </div>
-                        <a className="btn btn-sm btn-outline-primary" onClick={handleViewBoard}>
+                        <a className="btn btn-sm btn-outline-primary" href={"/board/" + props.board["id"]}>
                             View
                         </a>
                     </div>
                 ) : (
                     <div className="d-flex flex-wrap justify-content-end">
-                        <a className="btn btn-sm btn-outline-primary" onClick={handleViewBoard}>
+                        <a className="btn btn-sm btn-outline-primary" href={"/board/" + props.board["id"]}>
                             View
                         </a>
                     </div>
